add delete product

+++ backend/internal/models/models.go
@@ -29,9 +29,9 @@ type Product struct {
 	Description string `json:"description"`
 	Sizes     []string `json:"sizes,omitempty"`
 	Colors []int `json:"colors,omitempty"`
-	Weight float32 `json:"weight"`
-	Price float32 `json:"price"`
-	Rating float32 `json:"rating"`
+	Weight float64 `json:"weight"`
+	Price float64 `json:"price"`
+	Rating float64 `json:"rating"`
 	CategoryID int `json:"category_id"`
 }
 type Order struct {
